Move accessories folder under lib
