+++ lib/accessories/dimmer_switch.js
@@ -1,5 +1,5 @@
-const Device = require('../lib/device.js')
-const VivintDict = require("../lib/vivint_dictionary.json")
+const Device = require('../device.js')
+const VivintDict = require("../vivint_dictionary.json")
 const debounce = require("debounce-promise")
 
 class DimmerSwitch extends Device {
