+++ lib/accessories/lock.js
@@ -1,5 +1,5 @@
-const Device = require('../lib/device.js')
-const VivintDict = require("../lib/vivint_dictionary.json")
+const Device = require('../device.js')
+const VivintDict = require("../vivint_dictionary.json")
 
 class Lock extends Device {
     constructor(accessory, data, config, log, homebridge, vivintApi) {
