+++ lib/accessories/thermostat.js
@@ -1,6 +1,6 @@
-const Device = require('../lib/device.js')
-const VivintDict = require("../lib/vivint_dictionary.json")
-const ThermostatCharacteristicsModule = require("../lib/thermostat_characteristics.js")
+const Device = require('../device.js')
+const VivintDict = require("../vivint_dictionary.json")
+const ThermostatCharacteristicsModule = require("../thermostat_characteristics.js")
 const debounce = require("debounce-promise")
 
 class Thermostat extends Device {
