+++ lib/device.js
@@ -1,4 +1,4 @@
-const dataPatch = require("../lib/datapatch.js")
+const dataPatch = require("./datapatch.js")
 
 class Device {
     constructor(accessory, data, config, log, homebridge, vivintApi) {
