+++ lib/device_set.js
@@ -1,16 +1,16 @@
 const VivintDict = require("./vivint_dictionary.json")
 
-const ContactSensor = require("../accessories/contact_sensor.js")
-const SmokeSensor = require("../accessories/smoke_sensor.js")
-const CarbonMonoxideSensor = require("../accessories/carbon_monoxide_sensor.js")
-const MotionSensor = require("../accessories/motion_sensor.js")
-const Lock = require("../accessories/lock.js")
-const Thermostat = require("../accessories/thermostat.js")
-const GarageDoor = require("../accessories/garage_door.js")
-const Panel = require("../accessories/panel.js")
-const Camera = require("../accessories/camera.js")
-const LightSwitch = require("../accessories/light_switch.js")
-const DimmerSwitch = require("../accessories/dimmer_switch.js")
+const ContactSensor = require("./accessories/contact_sensor.js")
+const SmokeSensor = require("./accessories/smoke_sensor.js")
+const CarbonMonoxideSensor = require("./accessories/carbon_monoxide_sensor.js")
+const MotionSensor = require("./accessories/motion_sensor.js")
+const Lock = require("./accessories/lock.js")
+const Thermostat = require("./accessories/thermostat.js")
+const GarageDoor = require("./accessories/garage_door.js")
+const Panel = require("./accessories/panel.js")
+const Camera = require("./accessories/camera.js")
+const LightSwitch = require("./accessories/light_switch.js")
+const DimmerSwitch = require("./accessories/dimmer_switch.js")
 
 function DeviceSetModule(config, log, homebridge, vivintApi) {
   let PlatformAccessory = homebridge.platformAccessory
